requirements.txt
substrate-interface
python-dotenv